refactor(gateway): extract status-message helpers from run.py (slice 14 of #54962)
